+++ polaris.shopify.com/package.json
@@ -29,7 +29,8 @@
     "react-dom": "^17.0.2",
     "react-markdown": "^8.0.2",
     "use-dark-mode": "^2.3.1",
-    "remark-gfm": "^3.0.1"
+    "remark-gfm": "^3.0.1",
+    "lodash.throttle": "^4.1.1"
   },
   "devDependencies": {
     "@types/gtag.js": "^0.0.10",
@@ -38,6 +39,7 @@
     "@types/node": "17.0.21",
     "@types/prismjs": "^1.26.0",
     "@types/react": "*",
+    "@types/lodash.throttle": "^4.1.7",
     "eslint-config-next": "12.1.0",
     "eslint": "8.10.0",
     "execa": "^6.1.0",
